requirements.txt
behave
behave-django
django<1.11
django-constance
django-ipware
django-picklefield
django-statsd-mozilla
djangorestframework
gunicorn
raven
twisted
